Clear CodeQL high-severity alerts blocking releases (TASK-372/373) (#10)
* feat(ci/TASK-372): configure CodeQL paths-ignore for vendored deps

Adds .github/codeql/codeql-config.yml that excludes **/_deps/** and
.vscode/build*/** from CodeQL analysis, and wires the file into
.github/workflows/codeql-analysis.yml via config-file: on the
codeql-action/init step.

Removes 5 high-severity GHAS false-positive alerts on GoogleTest
source pulled into .vscode/build/_deps/ during the host-test build.

Also scaffolds TASK-373 (real cpp/comparison-with-wider-type fixes
in config_loader.cpp / config_loader_merge.cpp), which is the other
half required to clear the GHAS gate so the v0.5.0 release PR #9
can merge.


* fix(ble/TASK-373): widen profile-loop counters to size_t

Switches the JsonArray-iteration counter in
lib/PedalLogic/src/config_loader.cpp:154 (loadConfig) and
lib/PedalLogic/src/config_loader_merge.cpp:65 (mergeConfig) from
uint8_t to size_t. Both loops compared the narrow counter against
JsonArray::size() (size_t), which CodeQL flagged as
cpp/comparison-with-wider-type — the counter could in principle wrap
to zero and never terminate.

In practice the secondary guard `i < hardwareConfig.numProfiles` (where
numProfiles is uint8_t) bounds the counter at 255, so behaviour is
unchanged. config_loader.cpp adds a static_cast<uint8_t> at the
profileManager.addProfile() call site since the counter is now wider
than the API expects; the cast is safe by construction.

Clears CodeQL alerts #1 (config_loader.cpp) and #2
(config_loader_merge.cpp) so the v0.5.0 release PR #9 can pass the
GHAS code-scanning gate.


* close TASK-372: Configure CodeQL to ignore vendored test-framework code


* close TASK-373: Fix cpp/comparison-with-wider-type in config_loader


* chore(clang-tidy): cherry-pick brace fixes from TASK-371 to clear release-PR gate

Pulls in pedal_config.cpp and profile.cpp from feature/config-extensions
(squash branch / PR #9) to clear the pre-existing clang-tidy backlog on
main. Without this, PR #10 cannot pass the required clang-tidy CI gate
even though my own changes do not touch these files — the gate fails
on errors that already exist on main.

Mechanical fixes:
- braces around single-statement if/for bodies in pedal_config.cpp
  (8x) and profile.cpp (4x) — readability-braces-around-statements
- remove unused 'ILogger* createLogger()' forward declaration in
  pedal_config.cpp — readability-redundant-declaration

Identical to the corresponding hunks in TASK-371's squashed history;
no functional change. Once PR #10 merges, main is clang-tidy-clean
and PR #9 (v0.5.0 release) can rebase and merge without --admin
override.


---------

+++ lib/PedalLogic/src/config_loader.cpp
@@ -151,7 +151,7 @@ bool ConfigLoader::loadFromString(ProfileManager& profileManager,
     }
 
     JsonArray profiles = doc["profiles"];
-    for (uint8_t i = 0; i < profiles.size() && i < hardwareConfig.numProfiles; i++)
+    for (size_t i = 0; i < profiles.size() && i < hardwareConfig.numProfiles; i++)
     {
         JsonObject profileJson = profiles[i];
         const char* profileName = profileJson["name"] | "";
@@ -160,7 +160,7 @@ bool ConfigLoader::loadFromString(ProfileManager& profileManager,
         auto newProfile = std::make_unique<Profile>(profileName);
         newProfile->setDescription(profileDescription);
         populateProfileFromJson(*newProfile, profileJson["buttons"], keyboard);
-        profileManager.addProfile(i, std::move(newProfile));
+        profileManager.addProfile(static_cast<uint8_t>(i), std::move(newProfile));
     }
 
     profileManager.resetToFirstProfile();
